Start of /etc/lorax files.
The existing buildinstall system had all sorts of data that
would be more suited to configuration files stored directly
in the scripts.  The most difficult set to maintain was the
list of packages for install mode and rescue mode.

I have created a set of configuration files divided by
architecture that allow you to specify the packages to be
included in general or for specific platforms.