Add packages required for gtk3 and the new anaconda UI.
This doesn't get rid of the gtk2 stuff yet, though.  The intention here is
that you can use this lorax to generate an image containing either the old
anaconda or the newui branch, simply by including a different repo in your
tree composition kickstart file.

Also, it appears that some things in the tree still require gtk2 so we may
be stuck with both for the forseeable future.