Dynamically generate the list of installed platforms for .treeinfo
The list of ARM platforms was represented as a static list to be
installed in .treeinfo for Beaker support, but as ARM moves to use
the multiplatform kernel the platform specific kernel images will
no longer be needed.  This process is beginning in F18 (3.7 kernel)
with HighBank being the first to use the baseline kernel.  Due to
this change, there will be no 'highbank' platform images, but Beaker
tries to import all platforms listed in .treeinfo.  To avoid errors,
we should dynamically create the list of ARM plaforms, including
only those that are actually provided.

Signed-off-by: David A. Marlin <dmarlin@redhat.com>