Make sure grubby is installed for initrd creation (#1001896)
Some things need grubby. But they don't want to require them. So we make
sure it is available for other packages to use while building the
images.

Note that it is still removed in the cleanup script, and when installing
a system the grubby used will be the one pulled from the install repo.