Activate anaconda-shell@.service on switch to empty VT (#980062)
Right now, logind reserves tty6 for a login shell, which is not what we
want - normally anaconda puts Xorg there, and there's no need for a
login prompt anyway.

This configures logind to activate "anaconda-shell@.service" when a user
switches to an unused tty, and reserves tty2 for that purpose (which is
where users expect a shell anyway).

This will avoid us having login prompts that users don't know what to do
with. It also probably saves us a little bit of RAM.