runtime-install: Add rpm-ostree, move dnf here
rpm-ostree is a new payload type available in Anaconda, will be used
by Project Atomic (http://projectatomic.io) builds and the new Fedora
Cloud Docker Host.

dnf has previously existed, but as it's only required at installation
time, let's move it here as well.

Per discussion on the anaconda list, we don't want to pull these in as
a dependency of the "anaconda" package as initial-setup -> anaconda
which means they end up on every system.