add fedora-repos-anaconda to runtime environment
This is a virtual provides that should install the product specific yum
repo files in /etc/yum.repos.d/ with the one matching the product name
enabled. Eg. fedora-server.repo with an id of fedora-server and
enabled=1

Anaconda will then use this repository when 'Closest Mirror' is selected
as the installation source.

Signed-off-by: Brian C. Lane <bcl@redhat.com>