Revert "add fedora-repos-anaconda to runtime environment"
This reverts commit bc9b40f18f85f85aedea0a9a53a126815000c669.

anaconda-21.48.14-1 will revert to the below behavior, so remove the virtual
provide that was added for F21 Alpha/Beta.

You must include a repo in your installation environment that has an id
matching the lower case first part of your product name when split at the '-'
character.

If you don't do that closest mirror will not work.

For Fedora products this means including ONE repo with an id of
'fedora'.