livemedia-creator: Also generate an iPXE config
Currently lmc has some ad-hoc code to generate a `PXE_CONFIG`.
In practice we should probably split this out somehow, as in reality
downstream users need to edit it.  So it should be more explicitly
a template.

That aside, it turns out there's also iPXE which in addition to having
the critical `i` prefix, supports http and variable substitution in
the config file itself.

This makes it a lot more convenient to use - downstream users can more
easily edit the `set base-url` and have things work, it's easy to put
an iPXE config into a pastebin etc.