Add an option for disk_size in case kickstart parsing fails
In cases where the passed kickstart does not have any partition
sizes specified (autopart or similar), livemedia-creator defaults
to a 2MiB disk, which is too small for many use cases.

Add an option to specify the disk size in GiB in case users want
to override the default logic.

Forgot to make the type int