Fix aarch64 installs due to missing/unsupported packages
With commit fe17f97 changing the default from optional to required there's
a few packages that aren't currently supported on aarch64 that break the
compose. In particular aarch64 currently still doesn't have kexec, with luck
that might change in the F-25 cycle but until it does we need to have an
exception.

Signed-off-by: Peter Robinson <pbrobinson@fedoraproject.org>