Remove incorrect variables from rhel7-livemedia.ks example
The only environmental variable set during %post is $ANA_INSTALL_PATH
which points to the installroot (usually /mnt/sysimage). It can be used
in a %post --nochroot to copy things from the installer root to the
target's root filesystem.

This removes the %post --nochroot section because the image is not
accessable at this point in the process. Items that need to go into the
/ of the iso need to be added in the final iso creation templates, like
x86_64.tmpl

Resolves: rhbz#1430547