Add basic support for authorized API routes
A POST to /api/auth with "username" and "password" of a local user is
required. An access token will be returned, and it needs to be included
in an Authorization header for any requests to protected routes. eg.

Authorization: JWT <token>

This commit only protects /api/v0/recipes/list

It now requires python-jwt from EPEL, and python-flask-jwt which is
currently built in COPR from a branch of the upstream project.

https://copr.fedorainfracloud.org/coprs/g/weldr/python-flask-jwt/