Make authentication optional
If --noauth is passed to lorax-composer or if the config file has the
"auth" option set to "0" it will skip checking for JWT tokens when using
the @authentication() decorator on routes.

This is mostly needed so that the API will continue to work until the UI
and Cockpit add support for authentication.

It defaults to enabled, and and will only be disabled by a "0" setting,
every other value will result it in being enabled.