Test the API routes using authentication
The common API tests are in tests/pylorax/common.py, with __test__ set
to False so that nosetest won't try to test it without it being
subclassed.

The get, post, and delete methods are wrapped to include the JWT
Authorization header when it is set.

test_server.py runs the tests without any authentication.
test_server_auth.py run it with authentication, using a test method that
skips testing the local users.