Fix anaconda metapackage name
"anaconda-install-deps" was the original placeholder name
of the metapackage and it looks like I forgot to change it
to "anaconda-install-env-deps", which is the final name
we have decided sounds better.

Oops! (it's even correct in the commit message...)

(cherry picked from commit db9e2a1e41e70ef282cac0524dc004e9c1b59540)
Signed-off-by: Brian C. Lane <bcl@redhat.com>