templates: Stop using gconfset
We had only been indirectly pulling in GConf, and anyways
nothing was listening to these keys.

<kalev> I still think it's a fallout from https://src.fedoraproject.org/rpms/libgnomekbd/c/27a90d973faeddc74e6f78b149e480ea728fbe0e?branch=master

Really in general, if we wanted to make changes like this
it'd probably be a lot simpler to do them on boot or so.

https://bugzilla.redhat.com/show_bug.cgi?id=1581838
(cherry picked from commit bb3d8edd06eed7a28eadfc5bc583aae8d8af0b8c)